requirements.txt
blinker==1.9.0
charset-normalizer==3.4.4
click==8.3.1
et_xmlfile==2.0.0
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.2.6
opencv-python==4.12.0.88
openpyxl==3.1.5
pillow==12.1.0
pyzbar==0.1.9
qrcode==8.2
reportlab==4.4.7
Werkzeug==3.1.5
